CIV-3403
Changes for handling payment  success after initial failure
