+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/GeneralApplicationCreationNotificationHandlerTest.java
@@ -24,9 +24,7 @@ import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAUrgencyRequirement;
 import uk.gov.hmcts.reform.civil.sampledata.CallbackParamsBuilder;
 import uk.gov.hmcts.reform.civil.sampledata.CaseDataBuilder;
-import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
-import uk.gov.hmcts.reform.civil.service.NotificationService;
-import uk.gov.hmcts.reform.civil.service.SolicitorEmailValidation;
+import uk.gov.hmcts.reform.civil.service.*;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -34,12 +32,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
+import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_GENERAL_APPLICATION_RESPONDENT;
+import static uk.gov.hmcts.reform.civil.callback.CaseEvent.START_NOTIFICATION_PROCESS_MAKE_DECISION;
 import static uk.gov.hmcts.reform.civil.enums.BusinessProcessStatus.STARTED;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.NO;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.YES;
@@ -55,168 +54,34 @@ public class GeneralApplicationCreationNotificationHandlerTest extends BaseCallb
     private GeneralApplicationCreationNotificationHandler handler;
 
     @MockBean
-    private SolicitorEmailValidation solicitorEmailValidation;
+    private GeneralApplicationCreationNotificationService gaNotificationService;
 
-    @MockBean
-    private CaseDetailsConverter caseDetailsConverter;
 
-    @MockBean
-    private CoreCaseDataService coreCaseDataService;
+    private CallbackParams params;
 
-    @MockBean
-    private NotificationService notificationService;
-
-    @MockBean
-    private NotificationsProperties notificationsProperties;
-
-    private static final Long CASE_REFERENCE = 111111L;
-    private static final String PROCESS_INSTANCE_ID = "1";
-    private static final String DUMMY_EMAIL = "hmcts.civil@gmail.com";
-    private static final LocalDateTime DUMMY_DATE = LocalDateTime.of(2022, 02, 15, 12, 00, 00);
-    public static LocalDateTime NOTIFICATION_DEADLINE = LocalDateTime.of(2022, 02, 15, 12, 00, 00);
-
-    @Nested
-    class AboutToSubmitCallback {
-
-        @BeforeEach
-        void setup() {
-            when(notificationsProperties.getGeneralApplicationRespondentEmailTemplate())
-                .thenReturn("general-application-respondent-template-id");
-        }
-
-        @Test
-        void notificationShouldSendWhenInvoked() {
-            CaseData caseData = getCaseData(true);
-
-            when(solicitorEmailValidation
-                     .validateSolicitorEmail(any(), any()))
-                .thenReturn(caseData);
-
-            CallbackParams params = CallbackParamsBuilder.builder().of(ABOUT_TO_SUBMIT, caseData).request(
-                CallbackRequest.builder().eventId("NOTIFY_GENERAL_APPLICATION_RESPONDENT").build()).build();
-            handler.handle(params);
-
-            verify(notificationService, times(2)).sendMail(
-                DUMMY_EMAIL,
-                "general-application-respondent-template-id",
-                getNotificationDataMap(),
-                "general-application-respondent-notification-" + CASE_REFERENCE
-            );
-        }
-
-        @Test
-        void notificationShouldSendWhenInvokedTwice() {
-            CaseData caseData = getCaseData(true);
-
-            when(solicitorEmailValidation
-                     .validateSolicitorEmail(any(), any()))
-                .thenReturn(caseData);
-
-            CallbackParams params = CallbackParamsBuilder.builder().of(ABOUT_TO_SUBMIT, caseData).request(
-                CallbackRequest.builder().eventId("NOTIFY_GENERAL_APPLICATION_RESPONDENT").build()).build();
-            handler.handle(params);
-
-            verify(notificationService, times(2)).sendMail(
-                any(),
-                any(),
-                any(),
-                any()
-            );
-        }
-
-        @Test
-        void notificationShouldNotSendWhenInvokedWhenConditionsAreNotMet() {
-            CaseData caseData = getCaseData(false);
-
-            when(solicitorEmailValidation
-                     .validateSolicitorEmail(any(), any()))
-                .thenReturn(caseData);
-
-            CallbackParams params = CallbackParamsBuilder.builder().of(ABOUT_TO_SUBMIT, caseData).request(
-                CallbackRequest.builder().eventId("NOTIFY_GENERAL_APPLICATION_RESPONDENT").build()).build();
-            handler.handle(params);
-            verifyNoInteractions(notificationService);
-        }
-
-        private Map<String, String> getNotificationDataMap() {
-            return Map.of(
-                NotificationData.CASE_REFERENCE, CASE_REFERENCE.toString(),
-                NotificationData.APPLICANT_REFERENCE, "claimant",
-                NotificationData.GA_NOTIFICATION_DEADLINE,
-                NOTIFICATION_DEADLINE.format(DateTimeFormatter.ofPattern("dd MMMM yyyy"))
-            );
-        }
-
-        private CaseData getCaseData(boolean isMet) {
-
-            List<Element<GASolicitorDetailsGAspec>> respondentSols = new ArrayList<>();
-
-            GASolicitorDetailsGAspec respondent1 = GASolicitorDetailsGAspec.builder().id("id")
-                .email(DUMMY_EMAIL).organisationIdentifier("2").build();
-
-            GASolicitorDetailsGAspec respondent2 = GASolicitorDetailsGAspec.builder().id("id")
-                .email(DUMMY_EMAIL).organisationIdentifier("3").build();
-
-            respondentSols.add(element(respondent1));
-            respondentSols.add(element(respondent2));
-
-            if (isMet) {
-
-                return new CaseDataBuilder()
-                    .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id("id")
-                                                 .email(DUMMY_EMAIL).organisationIdentifier("1").build())
-                    .generalAppRespondentSolicitors(respondentSols)
-                    .businessProcess(BusinessProcess.builder().status(STARTED)
-                                         .processInstanceId(PROCESS_INSTANCE_ID).build())
-                    .gaInformOtherParty(GAInformOtherParty.builder().isWithNotice(YES).build())
-                    .gaUrgencyRequirement(GAUrgencyRequirement.builder().generalAppUrgency(NO).build())
-                    .parentClaimantIsApplicant(YES)
-                    .gaRespondentOrderAgreement(GARespondentOrderAgreement.builder().hasAgreed(NO).build())
-                    .respondentSolicitor1EmailAddress(DUMMY_EMAIL)
-                    .respondentSolicitor2EmailAddress(DUMMY_EMAIL)
-                    .applicant1OrganisationPolicy(OrganisationPolicy.builder()
-                                                      .organisation(Organisation.builder().organisationID("1").build())
-                                                      .build())
-                    .respondent1OrganisationPolicy(OrganisationPolicy.builder()
-                                                       .organisation(Organisation.builder().organisationID("2").build())
-                                                       .build())
-                    .respondent2OrganisationPolicy(OrganisationPolicy.builder()
-                                                       .organisation(Organisation.builder().organisationID("3").build())
-                                                       .build())
-                    .generalAppParentCaseLink(
-                        GeneralAppParentCaseLink
-                            .builder()
-                            .caseReference(CASE_REFERENCE.toString())
-                            .build())
-                    .generalAppDeadlineNotificationDate(DUMMY_DATE)
-                    .build();
-            } else {
-                return new CaseDataBuilder()
-                    .businessProcess(BusinessProcess.builder().status(STARTED)
-                                         .processInstanceId(PROCESS_INSTANCE_ID).build())
-                    .gaInformOtherParty(GAInformOtherParty.builder().isWithNotice(NO).build())
-                    .gaUrgencyRequirement(GAUrgencyRequirement.builder().generalAppUrgency(NO).build())
-                    .gaRespondentOrderAgreement(GARespondentOrderAgreement.builder().hasAgreed(NO).build())
-                    .ccdCaseReference(CASE_REFERENCE)
-                    .respondentSolicitor1EmailAddress(DUMMY_EMAIL)
-                    .respondentSolicitor2EmailAddress(DUMMY_EMAIL)
-                    .applicant1OrganisationPolicy(OrganisationPolicy.builder()
-                                                      .organisation(Organisation.builder().organisationID("1").build())
-                                                      .build())
-                    .respondent1OrganisationPolicy(OrganisationPolicy.builder()
-                                                       .organisation(Organisation.builder().organisationID("2").build())
-                                                       .build())
-                    .respondent2OrganisationPolicy(OrganisationPolicy.builder()
-                                                       .organisation(Organisation.builder().organisationID("3").build())
-                                                       .build())
-                    .generalAppParentCaseLink(
-                        GeneralAppParentCaseLink
-                            .builder()
-                            .caseReference(CASE_REFERENCE.toString())
-                            .build())
-                    .generalAppDeadlineNotificationDate(DUMMY_DATE)
-                    .build();
-            }
-        }
+    @Test
+    public void shouldReturnCorrectEvent() {
+        CaseData caseData = CaseDataBuilder.builder().judicialOrderMadeWithUncloakApplication().build();
+        params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
+        assertThat(handler.handledEvents()).contains(NOTIFY_GENERAL_APPLICATION_RESPONDENT);
     }
+
+    @Test
+    void shouldThrowException_whenNotificationSendingFails() {
+        var caseData = CaseDataBuilder.builder().generalOrderApplication()
+            .build();
+
+        doThrow(buildNotificationException())
+            .when(gaNotificationService)
+            .sendNotification(caseData);
+
+        params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
+        assertThrows(NotificationException.class, () -> handler.handle(params));
+    }
+
+    private NotificationException buildNotificationException() {
+        return new NotificationException(new Exception("Notification Exception"));
+    }
+
 }
+
