+++ src/test/java/uk/gov/hmcts/reform/civil/service/PaymentRequestUpdateCallbackServiceTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.civil.enums.CaseState;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.PaymentDetails;
@@ -29,6 +30,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.END_JUDGE_BUSINESS_PROCESS_GASPEC;
+import static uk.gov.hmcts.reform.civil.enums.CaseState.APPLICATION_PAYMENT_FAILED;
 import static uk.gov.hmcts.reform.civil.enums.CaseState.APPLICATION_ADD_PAYMENT;
 import static uk.gov.hmcts.reform.civil.enums.CaseState.PENDING_CASE_ISSUED;
 import static uk.gov.hmcts.reform.civil.enums.PaymentStatus.FAILED;
@@ -50,7 +52,8 @@ class PaymentRequestUpdateCallbackServiceTest {
     ObjectMapper objectMapper;
     @MockBean
     private CoreCaseDataService coreCaseDataService;
-
+    @MockBean
+    private GeneralApplicationCreationNotificationService gaNotificationService;
     @MockBean
     private JudicialNotificationService judicialNotificationService;
     @MockBean
@@ -204,5 +207,25 @@ class PaymentRequestUpdateCallbackServiceTest {
             .eventId(END_JUDGE_BUSINESS_PROCESS_GASPEC.name())
             .caseDetails(caseDetails)
             .build();
+    }
+    @Test
+    public void ShouldProceedAfterInitialPaymentFailureIsSuccess() {
+
+        CaseData caseData = CaseDataBuilder.builder().buildPaymentSuccesfulCaseData().toBuilder().build();
+        caseData = caseData.toBuilder().ccdState(APPLICATION_PAYMENT_FAILED).build();
+        CaseDetails caseDetails = buildCaseDetails(caseData);
+        when(coreCaseDataService.getCase(Long.valueOf(CASE_ID))).thenReturn(caseDetails);
+        when(caseDetailsConverter.toCaseData(caseDetails))
+            .thenReturn(caseData);
+        when(coreCaseDataService.startGaUpdate(any(), any())).thenReturn(startEventResponse(caseDetails));
+        when(coreCaseDataService.submitGaUpdate(any(), any())).thenReturn(caseData);
+
+        paymentRequestUpdateCallbackService.processCallback(buildServiceDto(PAID));
+        CaseState c = caseData.getCcdState();
+        verify(coreCaseDataService, times(1)).getCase(Long.valueOf(CASE_ID));
+        verify(coreCaseDataService, times(1)).startGaUpdate(any(), any());
+        verify(coreCaseDataService, times(1)).submitGaUpdate(any(), any());
+        verify(coreCaseDataService, times(1)).triggerEvent(any(), any());
+
     }
 }
